refactor : 중복클래스 제거 ( ladder, lines )
